Add config file and custom RuboCop rule
Introduce a new default configuration file for common settings. It sets the AllCops exclude paths and Rail's target version among others. For further details, comments with links have been included on the YML.

In addition, new custom rule `ModelCustomValidationMethodName` has been added under the Rails Cops module. This rule validates the naming of custom validation methods in models of a Rails application. Specifically, it checks if the custom validation method name includes a '_validation' suffix. The rule will output an offense if the naming does not conform to this standard. This is a necessary rule to keep method naming conventions consistent across the Rails project.