Refactor Rubocop config and add model_enum_setting
The Rubocop configuration file has been cleaned up. This refactor involves the removal of unnecessary exclusions and `inherit_mode`, a now obsolete attribute, is also deleted. Instead, there is now an inclusion for the `app/models` directory under the new `ModelEnumSetting` attribute under `Cops`.

To cater to a custom RuboCop rule for Rails model enums, a new cop called `ModelEnumSetting` has been introduced- it's committed to `lib/rails/cops.rb` and is defined at `lib/rails/cops/model_enum_setting.rb`.

The `ModelEnumSetting` will enforce a convention to use `_prefix` or `_suffix` with `enum`. If these are missing, it will provide a feedback message: "You are using enum without _prefix or _suffix, please add _prefix or _suffix to the enum". This is aimed at improving code readability and clarity about the role of the `enum` in the application.